+++ src/Symfony/Bridge/Doctrine/CacheWarmer/ProxyCacheWarmer.php
@@ -21,6 +21,8 @@ use Symfony\Component\HttpKernel\CacheWarmer\CacheWarmerInterface;
  * since this information is necessary to build the proxies in the first place.
  *
  * @author Benjamin Eberlei <kontakt@beberlei.de>
+ *
+ * @final since Symfony 7.1
  */
 class ProxyCacheWarmer implements CacheWarmerInterface
 {